Singleton-ify ErrorTransienceValue
We rely on ErrorTransienceValue being different across evaluations by
way of java reference inequality, instead make it a singleton and make
it so that ErrorTransienceValue compares equal to nothing, including itself.

Alternatively we could attach the corresponding Version to ErrorTransienceValue
and use that to force inequality across evaluations. However this gets into
the different possible Version implementations and their varying semantics, so
go with the simplest approach first and see how it works out.

--
MOS_MIGRATED_REVID=106001826