Cleanup ValueVisitor (and dirty QuiescingExecutor)
Raises the level of abstraction of ValueVisitor's dependence on
AbstractQueueVisitor. Except for the "ForTestingOnly" methods now
available on the QuiescingExecutor interface, ValueVisitor is
agnostic to the implementation of its executor.

This also cleans up the full spectrum of visibility modifiers on
ValueVisitor methods, all of which ought to be private.

--
MOS_MIGRATED_REVID=106847453