Speed up & deflake bazel_sandboxing_cpp_test (and probably other shell tests).
The "clean" command that is run after each shell test finishes (in the tear_down method of test-setup.sh) tries to remove all subdirectories of the output base. This includes the bazel-sandboxes directory. This, however is already being cleaned up continuously in the background by a thread pool setup in SandboxModule. By waiting for the pool to finish before the clean command runs, the race condition can be solved.

--
MOS_MIGRATED_REVID=107578302