Drop intermediate zip archive from py_binary output
If you feed a py_binary into a pkg_tar to a pkg_deb, you won't want to
have the deploy_zip intermediate artifact included in the resulting
package.  I'm think this change keeps it as an expected byproduct of
the build, but drops it from the list of files produced by the
y_binary rule.

Also trimmed a superfluous chmod from the final concat action.

--
Change-Id: Ib2bc05c2e43c7329bd9d92ea034f0c613f7fcbc6
Reviewed-on: https://bazel-review.googlesource.com/#/c/2280
MOS_MIGRATED_REVID=109290774