Scaffolding for implementing symlinks on Windows.
At first, this will only be used for emulating the planned implementation on Linux to validate it a little more before starting with the big work of porting everything to Windows in case it is doomed to failure.

In logging mode, the only places where we create symbolic links that we can't emulate with the plan (pointing to a non-existent file or to a file outside the output base and the source root, which are assumed to be writable):

- ExecutionTool.createOutputDirectoryLinks(). If we won't have the convenience symlinks on Windows, I won't shed a tear (I'm wondering why, though, because they are between the output base and the source tree)
- In the implementation of new_local_repository (Would need to be special-cased for Windows. No big deal.)
- In the implementation of the .tar.gz decompressor (doesn't seem to be serious, either.)

So this seems to be alright. Note, however, that we didn't check build-runfiles.cc, which might cause trouble. I don't remember any place where we create a link there that is illegal according to the above rules, though.

--
MOS_MIGRATED_REVID=112659070