RELNOTES[INC]: Remove JAVAC "Make" variable
The JAVAC "Make" variable is the javac packaged with the host JDK, which is not
necessarily the same version used to compile java_library rules.

Users should typically use java_library directly. If a manual javac invocation
is required, BazelJavac uses the same javac version and flags as java_library.

In the rare cases where the host JDK's compiler is actually desired, prefer
`$(JAVABASE)/bin/javac` to `$(JAVAC)` for clarity.

--
MOS_MIGRATED_REVID=113386241