Add escapes to quotation marks in Java version detection genrule.
Quotation mark character is significant for Win32's CreateProcess.
It should always be properly escaped when occuring inside arguments.

We might devise a common solution in GenRule for that, but for now
this change fixes the only genrule we use that runs into this issue.

Needed for #276.

--
MOS_MIGRATED_REVID=114881441