Remove current directory from PATH values.
Having an empty path component in the PATH is equivalent to having a path
component pointing to the current directory.  This is generally considered
harmful because it can lead to the execution of binaries planted in
specific locations without the user's knowledge (though I do not think there
is a problem per se in Bazel).

Fixes #958.

--
MOS_MIGRATED_REVID=115363638