Enable access to file values in repository context
ctx.path now works with labels. When given a label, ctx.path()
will returns the path to the corresponding file if the label
point to a file, or error out if it does not point to file.

Also fix a small case with repository_ctx.symlink that wasn't
creating its directories.

This is needed to enable reading template files to create a
template function (see http://goo.gl/fD4ZsY) for issue #893
(step 4 of http://goo.gl/OZV3o0).

--
MOS_MIGRATED_REVID=115438400