Canonicalize the name of the target itself for java indirect dep errors.
Roll-forward of I3762e6be911b79a133a2959fce3a8bf34d183232 with fixes.

Otherwise it prints the wrong message when the target which is missing
a dependency is in a repository other than the default one.

Second try with the new test disabled for jdk7 because that uses an old,
fixed version of the code.

--
Change-Id: I7e16fa874066e1b63064662df42a98784fcae233
Reviewed-on: https://bazel-review.googlesource.com/#/c/3032/
MOS_MIGRATED_REVID=116248259