Refactor spawnXcrunActionBuilder to take an explicit apple platform.
This both starts to clean up the code (making it clearer which platform each action is targeted to, caller by caller), as well as allows callers to issue
certain actions for platforms which are not implicitly ios.
We'll still need to migrate callers onto the new method signature.

--
MOS_MIGRATED_REVID=117982741