experimental UI: don't update progress bar after build for INFO
As explained in commit 37f3e9ee, after completion of the build, there
is no need any more to update the progress bar. Instead, its last line,
if any, can just float into the scroll-back buffer. From then on, STDOUT
and STDERR events are just passed through. Do so as well for any late
INFO, WARNING, or ERROR events. This will allow avoiding pointless
messages in commands like 'bazel clean' (the patch removing those messages
will also provide an appropriate integration test).

Unfortunately, to make this work properly, we must ensure that both
event handlers get informed about the fact that we have a no-build
command. To do this, we register the event handler with the intended
IO stream on the event bus, as well as a passively observing variant of the
event handler that has direct acces to the full output stream.

--
Change-Id: I148dc0065b6343e4a2a225b2d4fa615b4f720ff5
Reviewed-on: https://bazel-review.googlesource.com/#/c/3124
MOS_MIGRATED_REVID=118444275