Dig out the builtin include file from the filegroup of libc_top instead of special-casing it in CppConfiguration.
This seems to be the most reasonable solution. I was toying with the idea of adding a field to CROSSTOOL but that would fail if you set libc_top to something other than what was specified in that file. If I had a infinite amount of time, I'd create a custom rule called cc_libc where libc_top would point so that this file can be referenced by an attribute, but since I don't, this seems to be workable compromise.

Also note that contrary to what you'd glean from the code, we don't actually have "compile" and "link" filegroups for libc.

--
MOS_MIGRATED_REVID=118921101