Remove te protocol compiler rules from @bazel_tools.
The immmediate reason for this change is that we also need to add gRPC support to the proto rules, and we don't want to also support gRPC in a half-baked way.

This makes the Bazel binary much smaller and avoid giving false signals that we (for now) support protobuf compilation. The protobuf rules are only for compiling Bazel itself.


RELNOTES[INC]: Bazel does not embed protocol buffer-related rules anymore.

--
MOS_MIGRATED_REVID=119516246