Rollback of commit 6cb8d820689ad029a9d0dc4ee1100db9b2d96515.
*** Reason for rollback ***

Breaks ci.bazel.io

While the basics for fixing the build is easy (just a few typos in packages building), fixing the test is a bit more tricky. I see only one solution for fixing the test: use a select statement that would select the good bazel version but that would always pull JavaBuilder as an external dependency when we do test.

Better roll this back then check the JavaBuilder 0.1.0 as a binary in third_party before rolling forward (a similar change is still needed to decouple running the test and building the binary for JDK 7)

*** Original change description ***


Refactor build for JDK 7

Now the JDK 7 tuning happens all in Bazel, removing logic
from the CI script. It uses remote repositories to access
JDK 7 dependencies.

--
MOS_MIGRATED_REVID=119773123