Remove experimental_ui integration tests that depend on timing
While the timing assumptions in those tests were quite generous (a sleep
10 should not take 15s, an event supposed to happen every second should
be observed at least once in a 3 second interval), it seems that under
really heavily loaded test environments those assumptions do not hold
true. So remove the tests, as it is better to not have integration tests
for timing properties than to have flaky ones.

--
MOS_MIGRATED_REVID=119826405