Add the --grpc_port startup option and start a Java server if it's passed in.
Note that the presence of server/grpc_port does not guarantee that the server actually listens to it and we can't guarantee it, either, because it can always be kill -9'd.

I haven't decided yet how the transition between AF_UNIX and gRPC will work. For now, I'm happy that we can start up a Java server.

The way to get the kernel-chosen port is truly awful, but it is apparently impossible to do so in a different way:

https://github.com/grpc/grpc-java/issues/72

--
MOS_MIGRATED_REVID=119828354