4 of 5: Serialization of UnwrittenMergedAndroidData. Adding AndroidDataSerializer, the serialize_format proto, and KeyValueConsumers (utility class for keeping consumers straight).
The serializtion is a bit more manual as previous experience has proven to me that simply writing all the resources into a proto map and pulling them out is not performant in the least. So, the serializer stores each message independent, the keys and then the values allowing for potential lazy loading and other optimizations in the future.

Also adds tests for parsing and writing style resources.

--
MOS_MIGRATED_REVID=120274904