4.7 of 5: Remove serialization of transitive dependencies.
Serializer no longer serializes transitive resources. That turned out to be a performance issue with large trees. Instead,it relies on the build tool to indicate the direct and transitive resources. This reduces the number of keys that get pulled from diamond dependencies and stays much, much leaner.

--
MOS_MIGRATED_REVID=121382705